Move files around to make more sense
