+++ Argo/Functions/curry.swift
