+++ Argo/Functions/flatReduce.swift
