+++ Argo/Functions/sequence.swift
