+++ Argo/Operators/Operators.swift
@@ -1,4 +1,51 @@
+import Runes
+
 infix operator <| { associativity left precedence 150 }
 infix operator <|? { associativity left precedence 150 }
 infix operator <|| { associativity left precedence 150 }
 infix operator <||? { associativity left precedence 150 }
+
+// MARK: Values
+
+// Pull value from JSON
+public func <|<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, key: String) -> Decoded<A> {
+  return decodeObject(json) >>- { (A.decode <^> $0[key]) ?? .MissingKey(key) }
+}
+
+// Pull optional value from JSON
+public func <|?<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, key: String) -> Decoded<A?> {
+  return .optional(json <| key)
+}
+
+// Pull embedded value from JSON
+public func <|<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, keys: [String]) -> Decoded<A> {
+  return flatReduce(keys, json, <|) >>- A.decode
+}
+
+// Pull embedded optional value from JSON
+public func <|?<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, keys: [String]) -> Decoded<A?> {
+  return .optional(json <| keys)
+}
+
+// MARK: Arrays
+
+// Pull array from JSON
+public func <||<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, key: String) -> Decoded<[A]> {
+  return json <| key >>- decodeArray
+}
+
+// Pull optional array from JSON
+public func <||?<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, key: String) -> Decoded<[A]?> {
+  return .optional(json <|| key)
+}
+
+// Pull embedded array from JSON
+public func <||<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, keys: [String]) -> Decoded<[A]> {
+  return json <| keys >>- decodeArray
+}
+
+// Pull embedded optional array from JSON
+public func <||?<A where A: JSONDecodable, A == A.DecodedType>(json: JSON, keys: [String]) -> Decoded<[A]?> {
+  return .optional(json <|| keys)
+}
+
